Implement isApplicable behavior to match the new SPI added in branch-api-plugin
This new SPI allows the creation of combination strategies which can support an
AND logic behavior of the installed strategies. By separating the isAutomaticBuild
and isApplicable responsibilities the calling code can safely apply a combiniation
logic of automatic unless denied logic that makes the result correct for each kind
of build request.